app/assignment3/requirements.txt

torch
torchvision
fastapi
uvicorn
pillow
